Normalize parameters names like `_$N` to `_`
TruffleRuby adds a "$N" suffix to parameter names to make them unique,
but that is only possible if the parameter name is `_`. This internal implementation detail then leaks to the user when the user calls `parameters` on a method or proc that has multiple parameters
defined with the name `_`.

This commit matches parameter names that start with `_$` and normalizes
them to be `_` as they are being assembled into the `parameters` array.